FAT-5137/C376606 (#1663)
* added test C376606

* small change
