+++ cypress/e2e/marc-authority/reporting/marc-authority-reporting.cy.js
@@ -1,21 +1,21 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import TestTypes from '../../support/dictionary/testTypes';
-import DevTeams from '../../support/dictionary/devTeams';
-import Permissions from '../../support/dictionary/permissions';
-import TopMenu from '../../support/fragments/topMenu';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import MarcAuthority from '../../support/fragments/marcAuthority/marcAuthority';
-import Users from '../../support/fragments/users/users';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import MarcAuthorities from '../../support/fragments/marcAuthority/marcAuthorities';
-import QuickMarcEditor from '../../support/fragments/quickMarcEditor';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import InventoryInstances from '../../support/fragments/inventory/inventoryInstances';
-import DateTools from '../../support/utils/dateTools';
-import ExportManagerSearchPane from '../../support/fragments/exportManager/exportManagerSearchPane';
-import FileManager from '../../support/utils/fileManager';
-import Parallelization from '../../support/dictionary/parallelization';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import TestTypes from '../../../support/dictionary/testTypes';
+import DevTeams from '../../../support/dictionary/devTeams';
+import Permissions from '../../../support/dictionary/permissions';
+import TopMenu from '../../../support/fragments/topMenu';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import MarcAuthority from '../../../support/fragments/marcAuthority/marcAuthority';
+import Users from '../../../support/fragments/users/users';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import MarcAuthorities from '../../../support/fragments/marcAuthority/marcAuthorities';
+import QuickMarcEditor from '../../../support/fragments/quickMarcEditor';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import InventoryInstances from '../../../support/fragments/inventory/inventoryInstances';
+import DateTools from '../../../support/utils/dateTools';
+import ExportManagerSearchPane from '../../../support/fragments/exportManager/exportManagerSearchPane';
+import FileManager from '../../../support/utils/fileManager';
+import Parallelization from '../../../support/dictionary/parallelization';
 
 describe('MARC Authority -> Reporting | MARC authority', () => {
   const testData = {
